requirements.txt
# Django & DRF
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0

# Database
psycopg2-binary==2.9.9

# CORS
django-cors-headers==4.3.0

# Environment variables
python-decouple==3.8

# Excel processing
pandas==2.1.3
openpyxl==3.1.2

# Celery (optional for async tasks)
celery==5.3.4
redis==5.0.1

# Django filters
django-filter==23.5

# API documentation (optional)
drf-yasg==1.21.7

# Production server
gunicorn==21.2.0

# Testing
pytest==7.4.3
pytest-django==4.7.0